Finale

+++ app/Http/Controllers/ChatRoomsController.php
@@ -11,21 +11,28 @@ class ChatRoomsController extends Controller
     {
         $result = $request->validate([
             'name' => 'required|string',
-            'user_id' => 'required|integer',
         ]);
 
-        $chatRoom = Chatroom::create($result);
+        $result['user_id'] = auth()->user()->id;
+
+        Chatroom::create($result);
         return redirect()->route('home');
     }
 
     public function destroy(ChatRoom $chatRoom)
     {
+        if($chatRoom->userId != auth()->id()) {
+            return;
+        }
         $chatRoom->delete();
         return redirect()->route('home');
     }
 
     public function update(ChatRoom $chatRoom, Request $request)
     {
+        if($chatRoom->userId != auth()->id()) {
+            return;
+        }
         $chatRoom->update(['name' => $request->name]);
         return back();
     }
